+++ src/migrations/1660056506587_collections.sql
@@ -1,0 +1,4 @@
+-- Up Migration
+ALTER TABLE "collections" ADD COLUMN "non_flagged_token_set_id" TEXT;
+
+-- Down Migration